Update .npmignore to ship source map
In the published version of this package, the index.js includes a source map comment, https://unpkg.com/browse/graphql-fields-list@2.1.3/index.js :
```
//# sourceMappingURL=index.js.map
```
But this source map is not included, which can cause issues when trying to debug with source maps.

To fix this can you either turn of source maps, thus removing the comment in the published version or as proposed include the source map and source in the published package.